# openapi.yaml
openapi: 3.0.3
info:
  title: Lightning Engine API
  description: |
    REST API for the Lightning Engine ECS game server.

    ## Overview

    The engine uses a tick-based simulation model:
    1. **Commands** are queued via POST requests or WebSocket
    2. **Tick** advances process commands and run all module systems
    3. **Snapshots** return the current ECS state

    ## WebSocket Endpoints

    ### Snapshot Streaming
    Connect via WebSocket at `ws://localhost:8080/ws/containers/{containerId}/snapshots/{matchId}` for real-time updates.

    ### Command WebSocket (Recommended for high-performance)
    Connect via WebSocket at `ws://localhost:8080/containers/{containerId}/commands?token=xxx` to submit commands using Protocol Buffers.

    **Authentication:** JWT token passed as query parameter (must have `admin` or `command_manager` role).

    **Protocol:** Binary Protocol Buffer messages (`CommandRequest`/`CommandResponse`).

    See `api-proto/src/main/proto/command.proto` for message definitions.

  version: 1.0.0
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/snapshots/match/{matchId}:
    get:
      tags: [Snapshots]
      summary: Get snapshot for match
      description: Returns current ECS state filtered to a specific match
      operationId: getMatchSnapshot
      parameters:
        - $ref: '#/components/parameters/matchId'
      responses:
        '200':
          description: Match snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotResponse'
              example:
                matchId: 1
                tick: 42
                data:
                  MoveModule:
                    POSITION_X: [100.0, 200.0, 150.0]
                    POSITION_Y: [50.0, 75.0, 100.0]
                    VELOCITY_X: [1.0, -1.0, 0.0]
                  SpawnModule:
                    ENTITY_TYPE: [1.0, 2.0, 1.0]
                    OWNER_ID: [1.0, 1.0, 2.0]
        '404':
          description: Match not found

  # Module endpoints
...
  /api/resources/{resourceId}:
    get:
      tags: [Resources]
      summary: Get resource metadata
      operationId: getResource
      parameters:
        - $ref: '#/components/parameters/resourceId'
      responses:
        '200':
          description: Resource metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceResponse'
        '404':
          description: Resource not found
    delete:
      tags: [Resources]
      summary: Delete resource
      operationId: deleteResource
      parameters:
        - $ref: '#/components/parameters/resourceId'
      responses:
        '204':
          description: Resource deleted
        '404':
          description: Resource not found
...
components:
  parameters:
    matchId:
      name: matchId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    resourceId:
      name: resourceId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
...
    SnapshotResponse:
      type: object
      properties:
        matchId:
          type: integer
          format: int64
        tick:
          type: integer
          format: int64
        data:
          type: object
          additionalProperties:
            type: object
            additionalProperties:
              type: array
              items:
                type: number
                format: float
          description: |
            Columnar ECS data organized by module.
            Structure: { "ModuleName": { "COMPONENT_NAME": [value1, value2, ...] } }
            Each array contains one value per entity, in entity order.
...
    ResourceResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        type:
          type: string
          enum: [TEXTURE]
        size:
          type: integer
          description: Size in bytes
...